+++ Common/src/main/java/fuzs/puzzleslib/api/capability/v3/data/SyncStrategy.java
@@ -1,8 +1,7 @@
 package fuzs.puzzleslib.api.capability.v3.data;
 
 import fuzs.puzzleslib.impl.PuzzlesLibMod;
-import fuzs.puzzleslib.impl.capability.ClientboundSyncCapabilityMessage;
-import net.minecraft.resources.ResourceLocation;
+import fuzs.puzzleslib.impl.capability.ClientboundEntityCapabilityMessage;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 
@@ -16,7 +15,7 @@ public enum SyncStrategy {
     MANUAL {
 
         @Override
-        public <T extends Entity> void send(ResourceLocation identifier, CapabilityComponent<T> capabilityComponent) {
+        public void send(Entity entity, ClientboundEntityCapabilityMessage message) {
 
         }
     },
@@ -28,9 +27,9 @@ public enum SyncStrategy {
     SELF {
 
         @Override
-        public <T extends Entity> void send(ResourceLocation identifier, CapabilityComponent<T> capabilityComponent) {
-            if (capabilityComponent.getHolder() instanceof ServerPlayer player) {
-                PuzzlesLibMod.NETWORK.sendTo(player, this.toPacket(identifier, capabilityComponent));
+        public void send(Entity entity, ClientboundEntityCapabilityMessage message) {
+            if (entity instanceof ServerPlayer player) {
+                PuzzlesLibMod.NETWORK.sendTo(player, message);
             }
         }
     },
@@ -42,21 +41,16 @@ public enum SyncStrategy {
     TRACKING {
 
         @Override
-        public <T extends Entity> void send(ResourceLocation identifier, CapabilityComponent<T> capabilityComponent) {
-            PuzzlesLibMod.NETWORK.sendToAllTracking(capabilityComponent.getHolder(), this.toPacket(identifier, capabilityComponent), true);
+        public void send(Entity entity, ClientboundEntityCapabilityMessage message) {
+            PuzzlesLibMod.NETWORK.sendToAllTracking(entity, message, true);
         }
     };
 
     /**
      * Send a {@link CapabilityComponent} to remotes.
      *
-     * @param identifier          identifier from {@link CapabilityKey#identifier()}
-     * @param capabilityComponent the component to send to remotes
-     * @param <T>                 entity type
+     * @param entity  the capability component holder
+     * @param message the message to send
      */
-    public abstract <T extends Entity> void send(ResourceLocation identifier, CapabilityComponent<T> capabilityComponent);
-
-    final <T extends Entity> ClientboundSyncCapabilityMessage toPacket(ResourceLocation identifier, CapabilityComponent<T> capabilityComponent) {
-        return new ClientboundSyncCapabilityMessage(identifier, capabilityComponent.getHolder().getId(), capabilityComponent.toCompoundTag());
-    }
+    public abstract void send(Entity entity, ClientboundEntityCapabilityMessage message);
 }