+++ Common/src/main/java/fuzs/puzzleslib/impl/PuzzlesLibMod.java
@@ -3,7 +3,7 @@ package fuzs.puzzleslib.impl;
 import fuzs.puzzleslib.api.core.v1.CommonAbstractions;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.api.network.v3.NetworkHandlerV3;
-import fuzs.puzzleslib.impl.capability.ClientboundSyncCapabilityMessage;
+import fuzs.puzzleslib.impl.capability.ClientboundEntityCapabilityMessage;
 import fuzs.puzzleslib.impl.core.ClientboundModListMessage;
 import fuzs.puzzleslib.impl.core.EventHandlerProvider;
 import fuzs.puzzleslib.impl.core.ModContext;
@@ -17,7 +17,7 @@ public class PuzzlesLibMod extends PuzzlesLib implements ModConstructor {
     public static final NetworkHandlerV3 NETWORK = NetworkHandlerV3.builder(MOD_ID)
             .registerSerializer(ClientboundAddEntityPacket.class, (friendlyByteBuf, clientboundAddEntityPacket) -> clientboundAddEntityPacket.write(friendlyByteBuf), ClientboundAddEntityPacket::new)
             .optional()
-            .registerClientbound(ClientboundSyncCapabilityMessage.class)
+            .registerClientbound(ClientboundEntityCapabilityMessage.class)
             .registerClientbound(ClientboundAddEntityDataMessage.class)
             .registerClientbound(ClientboundModListMessage.class);