+++ Common/src/main/java/fuzs/puzzleslib/impl/capability/ClientboundEntityCapabilityMessage.java
@@ -12,14 +12,14 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
 
-public record ClientboundSyncCapabilityMessage(ResourceLocation identifier, int entityId, CompoundTag tag) implements ClientboundMessage<ClientboundSyncCapabilityMessage> {
+public record ClientboundEntityCapabilityMessage(ResourceLocation identifier, int entityId, CompoundTag tag) implements ClientboundMessage<ClientboundEntityCapabilityMessage> {
 
     @Override
-    public ClientMessageListener<ClientboundSyncCapabilityMessage> getHandler() {
+    public ClientMessageListener<ClientboundEntityCapabilityMessage> getHandler() {
         return new ClientMessageListener<>() {
 
             @Override
-            public void handle(ClientboundSyncCapabilityMessage message, Minecraft client, ClientPacketListener handler, LocalPlayer player, ClientLevel level) {
+            public void handle(ClientboundEntityCapabilityMessage message, Minecraft client, ClientPacketListener handler, LocalPlayer player, ClientLevel level) {
                 Entity entity = level.getEntity(message.entityId);
                 if (entity != null) {
                     CapabilityKey<?, ?> capabilityKey = GlobalCapabilityRegister.retrieve(message.identifier);