+++ NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/capability/data/NeoForgeEntityCapabilityKey.java
@@ -6,6 +6,9 @@ import fuzs.puzzleslib.api.capability.v3.data.EntityCapabilityKey;
 import fuzs.puzzleslib.api.capability.v3.data.SyncStrategy;
 import net.minecraft.world.entity.Entity;
 import net.neoforged.neoforge.attachment.AttachmentType;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.entity.living.LivingConversionEvent;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 import net.neoforged.neoforge.registries.DeferredHolder;
 
 import java.util.function.Predicate;
@@ -24,6 +27,13 @@ public class NeoForgeEntityCapabilityKey<T extends Entity, C extends CapabilityC
             throw new IllegalStateException("Sync strategy has already been set!");
         } else {
             this.syncStrategy = syncStrategy;
+            if (this.syncStrategy != SyncStrategy.MANUAL) {
+                NeoForge.EVENT_BUS.addListener(this::onPlayerLoggedIn);
+                NeoForge.EVENT_BUS.addListener(this::onPlayerChangedDimension);
+                if (syncStrategy == SyncStrategy.TRACKING) {
+                    NeoForge.EVENT_BUS.addListener(this::onStartTracking);
+                }
+            }
             return this;
         }
     }
@@ -34,6 +44,10 @@ public class NeoForgeEntityCapabilityKey<T extends Entity, C extends CapabilityC
             throw new IllegalStateException("Copy strategy has already been set!");
         } else {
             this.copyStrategy = copyStrategy;
+            if (this.copyStrategy != CopyStrategy.NEVER) {
+                NeoForge.EVENT_BUS.addListener(this::onPlayerClone);
+                NeoForge.EVENT_BUS.addListener(this::onAfterLivingConversion);
+            }
             return this;
         }
     }
@@ -46,5 +60,36 @@ public class NeoForgeEntityCapabilityKey<T extends Entity, C extends CapabilityC
     @Override
     public CopyStrategy getCopyStrategy() {
         return this.copyStrategy;
+    }
+
+    private void onPlayerClone(final PlayerEvent.Clone evt) {
+        if (evt.isWasDeath() && this.isProvidedBy(evt.getOriginal()) && this.isProvidedBy(evt.getEntity())) {
+            this.copyStrategy.copy(evt.getOriginal(), this.get((T) evt.getOriginal()), evt.getEntity(), this.get((T) evt.getEntity()));
+        }
+    }
+
+    private void onAfterLivingConversion(final LivingConversionEvent.Post evt) {
+        if (this.isProvidedBy(evt.getEntity()) && this.isProvidedBy(evt.getOutcome())) {
+            this.copyStrategy.copy(evt.getEntity(), this.get((T) evt.getEntity()), evt.getOutcome(), this.get((T) evt.getOutcome()));
+        }
+    }
+
+    private void onPlayerLoggedIn(final PlayerEvent.PlayerLoggedInEvent evt) {
+        if (this.isProvidedBy(evt.getEntity())) {
+            this.syncStrategy.send(evt.getEntity(), this.toPacket(this.get((T) evt.getEntity())));
+        }
+    }
+
+    private void onPlayerChangedDimension(final PlayerEvent.PlayerChangedDimensionEvent evt) {
+        if (this.isProvidedBy(evt.getEntity())) {
+            this.syncStrategy.send(evt.getEntity(), this.toPacket(this.get((T) evt.getEntity())));
+        }
+    }
+
+    private void onStartTracking(final PlayerEvent.StartTracking evt) {
+        if (this.isProvidedBy(evt.getTarget()) && this.syncStrategy == SyncStrategy.TRACKING) {
+            // we only want to sync to the client that just started tracking, so use SyncStrategy#SELF
+            SyncStrategy.SELF.send(evt.getEntity(), this.toPacket(this.get((T) evt.getTarget())));
+        }
     }
 }