+++ CHANGELOG.md
@@ -6,5 +6,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v20.4.0-1.20.4] - 2024-01-11
 - Ported to Minecraft 1.20.4
-### Added
-- Added support for NeoForge
+- Ported to NeoForge